Make payload handling an optional feature
