+++ Cargo.toml
@@ -27,6 +27,21 @@ required-features = ["test-with-podman"]
 path = "tests/compat.rs"
 
 [[test]]
+name = "building"
+required-features = ["payload"]
+path = "tests/building.rs"
+
+[[test]]
+name = "compare"
+required-features = ["payload"]
+path = "tests/compare.rs"
+
+[[test]]
+name = "payload"
+required-features = ["payload"]
+path = "tests/payload.rs"
+
+[[test]]
 name = "signatures"
 required-features = ["signature-meta"]
 path = "tests/signatures.rs"
@@ -71,10 +86,13 @@ pgp = "0.19.0"
 [features]
 default = [
     "signature-pgp",
+    "payload",
     "gzip-compression",
     "zstd-compression",
     "xz-compression",
 ]
+
+payload = []
 
 gzip-compression = ["flate2"]
 zstd-compression = ["zstd"]
@@ -85,10 +103,42 @@ signature-pgp = ["signature-meta", "pgp", "chrono", "getrandom"]
 signature-meta = []
 
 # Segregate tests that require podman to be installed
-test-with-podman = ["signature-pgp"]
+test-with-podman = ["signature-pgp", "payload"]
 
 zstdmt = ["zstd-compression", "zstd/zstdmt"]
 
 [[example]]
 name = "hsm-signing"
 required-features = ["signature-pgp"]
+
+[[example]]
+name = "parse-repo"
+
+[[example]]
+name = "inspect-rpm"
+
+[[example]]
+name = "verify-signatures"
+required-features = ["signature-pgp"]
+
+[[example]]
+name = "build-rpm"
+required-features = ["payload"]
+
+[[example]]
+name = "compare-versions"
+
+[[example]]
+name = "extract-rpm"
+required-features = ["payload"]
+
+[[example]]
+name = "sign-rpm"
+required-features = ["signature-pgp"]
+
+[[example]]
+name = "remote-signing"
+required-features = ["signature-pgp"]
+
+[[example]]
+name = "diff-rpms"
