+++ src/rpm/builder.rs
@@ -15,6 +15,7 @@ use digest::Digest;
 use super::compressor::Compressor;
 use super::headers::*;
 use super::payload;
+use super::util::{ChecksummingWriter, HashKind};
 use crate::errors::*;
 use crate::{Evr, Timestamp, constants::*};
 
@@ -24,13 +25,7 @@ use crate::signature;
 use crate::Package;
 use crate::PackageMetadata;
 
-use crate::{CompressionType, CompressionWithLevel};
-
-#[derive(Copy, Clone, PartialEq)]
-pub enum RpmFormat {
-    V4,
-    V6,
-}
+use crate::{CompressionType, CompressionWithLevel, RpmFormat};
 
 #[derive(Copy, Clone, PartialEq)]
 pub struct BuildConfig {
