+++ src/rpm/mod.rs
@@ -1,10 +1,15 @@
+#[cfg(feature = "payload")]
 mod builder;
 mod compressor;
+#[cfg(feature = "payload")]
+mod content;
 mod filecaps;
 mod headers;
 mod package;
+#[cfg(feature = "payload")]
 mod payload;
 mod timestamp;
+#[cfg(feature = "payload")]
 mod util;
 
 pub mod signature;
@@ -15,7 +20,11 @@ pub use compressor::*;
 
 pub use package::*;
 
+#[cfg(feature = "payload")]
 pub use builder::*;
+
+#[cfg(feature = "payload")]
+pub use content::*;
 
 pub use timestamp::*;
